fix(model): 수익률 계산 오류 및 로직 개선

+++ src/main/java/lotto/model/LottoResult.java
@@ -85,8 +85,8 @@ public class LottoResult {
                 winningMoney += winningPrize[4];
                 resultList[4]++;
             }
-            rateOfReturn = winningMoney/purchaseMoney*100;
-            rateOfReturn = Math.round(rateOfReturn*100.0)/100.0;
+            rateOfReturn = (double)winningMoney / purchaseMoney * 100;
+            rateOfReturn = Math.round(rateOfReturn * 100.0) / 100.0;
         }
         return resultList;
     }
